refactor(data): split ai_parameter_registry under non-comment line-size gate (Refs #3794)
The #3794 victory-config additions pushed ai_parameter_registry.dart to 1021
non-comment lines, failing repo.dart_file_non_comment_line_size (<=1000) and
blocking the Quality check. Extract the shared AiParameter/AiParameterCategory
types into ai_parameter.dart and the victory-config parameter list (with its
_vcInt/_vcDouble helpers) into ai_parameter_victory_config_params.dart as
standalone libraries (no part fragments, per SPEC). The registry re-exports
ai_parameter.dart so the public API is unchanged.
